Augment bad-words.d.ts declaration to be natively supported by typescript instead of manually enforcing the type using "as" keyword.
